+++ Visual Studio Project/ZWaveJS.NET/ZWaveNode.cs
@@ -49,6 +49,13 @@ namespace ZWaveJS.NET
             Notification?.Invoke(this, CCID, Args);
         }
 
+        public delegate void NodeDeadEvent(ZWaveNode Node);
+        public event NodeDeadEvent NodeDead;
+        internal void Trigger_NodeDead()
+        {
+            NodeDead?.Invoke(this);
+        }
+
         public delegate void AwakeEvent(ZWaveNode Node);
         public event AwakeEvent NodeAwake;
         internal void Trigger_NodeAwake()
@@ -105,7 +112,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.AbortFirmwareUpdate);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
           
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
@@ -130,7 +137,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.BeginFirmwareUpdate);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("firmwareFile", FileData);
             Request.Add("firmwareFilename", FI.Name);
 
@@ -153,7 +160,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.RefreshInfo);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -175,7 +182,7 @@ namespace ZWaveJS.NET
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetValue);
             Request.Add("valueId", ValueID);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
             Driver.Client.Send(RequestPL);
@@ -196,7 +203,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.SetValue);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", ValueID);
             Request.Add("value", Value);
 
@@ -224,7 +231,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.PollValue);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", ValueID);
 
             string RequestPL = Newtonsoft.Json.JsonConvert.SerializeObject(Request);
@@ -246,7 +253,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetDefinedValueIDs);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
 
 
             string RequestPL = JsonConvert.SerializeObject(Request);
@@ -268,7 +275,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.GetValueMetadata);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("valueId", VID);
 
 
@@ -291,7 +298,7 @@ namespace ZWaveJS.NET
             Dictionary<string, object> Request = new Dictionary<string, object>();
             Request.Add("messageId", ID);
             Request.Add("command", Enums.Commands.InvokeCCAPI);
-            Request.Add("nodeId", this.nodeId);
+            Request.Add("nodeId", this.id);
             Request.Add("commandClass", CommandClass);
             Request.Add("methodName", Method);
             Request.Add("args", Params);
@@ -309,17 +316,82 @@ namespace ZWaveJS.NET
             return EP;
         }
 
+        public Endpoint[] GetAllEndpoints()
+        {
+            return endpoints;
+        }
+
+        public Task<int> GetEndpointCount()
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<int> Result = new TaskCompletionSource<int>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(JO.SelectToken("result.count").Value<int>()) ;
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.GetEndpointCount);
+            Request.Add("nodeId", this.id);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<Enums.SecurityClass> GetHighestSecurityClass()
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<Enums.SecurityClass> Result = new TaskCompletionSource<Enums.SecurityClass>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                int Value = JO.SelectToken("result.highestSecurityClass").Value<int>();
+                Result.SetResult((Enums.SecurityClass)Value);
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.GetHighestSecurityClass);
+            Request.Add("nodeId", this.id);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
+        public Task<bool> HasSecurityClass(Enums.SecurityClass Class)
+        {
+            Guid ID = Guid.NewGuid();
+
+            TaskCompletionSource<bool> Result = new TaskCompletionSource<bool>();
+
+            Driver.Callbacks.Add(ID, (JO) =>
+            {
+                Result.SetResult(JO.SelectToken("result.hasSecurityClass").Value<bool>());
+            });
+
+            Dictionary<string, object> Request = new Dictionary<string, object>();
+            Request.Add("messageId", ID);
+            Request.Add("command", Enums.Commands.HasSecurityClass);
+            Request.Add("nodeId", this.id);
+            Request.Add("securityClass", Class);
+
+            string RequestPL = JsonConvert.SerializeObject(Request);
+            Driver.Client.Send(RequestPL);
+
+            return Result.Task;
+        }
+
         [Newtonsoft.Json.JsonProperty]
         internal Endpoint[] endpoints { get; set; }
 
-        [Newtonsoft.Json.JsonProperty]
-        public int nodeId { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int index { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int installerIcon { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int userIcon { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public Enums.NodeStatus status { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
@@ -343,10 +415,6 @@ namespace ZWaveJS.NET
         [Newtonsoft.Json.JsonProperty]
         public DeviceConfig deviceConfig { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public string label { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
-        public int interviewAttempts { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public object isFrequentListening { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public long maxDataRate { get; internal set; }
@@ -359,23 +427,24 @@ namespace ZWaveJS.NET
         [Newtonsoft.Json.JsonProperty]
         public bool supportsSecurity { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public int nodeType { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public int zwavePlusNodeType { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public int zwavePlusRoleType { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public DeviceClass deviceClass { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public CommandClass[] commandClasses { get; internal set; }
-        [Newtonsoft.Json.JsonProperty]
         public string interviewStage { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
         public string deviceDatabaseUrl { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public int highestSecurityClass { get; internal set; }
+        public int interviewAttempts { get; internal set; }
         [Newtonsoft.Json.JsonProperty]
-        public ValueDump[] values { get; internal set; }
+        public string label { get; internal set; }
+        [Newtonsoft.Json.JsonProperty]
+        public int nodeType { get; internal set; }
+
+        [Newtonsoft.Json.JsonProperty(PropertyName = "nodeId")]
+        public int id { get; internal set; }
 
         private bool _KeepAwake;
         [Newtonsoft.Json.JsonProperty]
@@ -395,7 +464,7 @@ namespace ZWaveJS.NET
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.KeepNodeAwake);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("keepAwake", _KeepAwake);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);
@@ -422,7 +491,7 @@ namespace ZWaveJS.NET
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.SetName);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("name", _Name);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);
@@ -450,7 +519,7 @@ namespace ZWaveJS.NET
                     Dictionary<string, object> Request = new Dictionary<string, object>();
                     Request.Add("messageId", ID);
                     Request.Add("command", Enums.Commands.SetLocation);
-                    Request.Add("nodeId", this.nodeId);
+                    Request.Add("nodeId", this.id);
                     Request.Add("location", _Location);
 
                     string RequestPL = JsonConvert.SerializeObject(Request);